First Game Controller support (not finished yet)
With this commit it becomes possible to use a Game Controller and
nagivate through the menus of the game (there is no graphical
feedback yet). This support is incomplete and does not support
duels yet...

This commit also adds a new dependency to the project: SDL2

if you want to compile under Windows you will need to modify the
premake4.lua to support SDL2.