Improve wildcard handling
Can now detect multiple instances of character immediately after wildcard.

For example, Dinomist Plesios has "Dinomist" in both its Pendulum Effect and Monster effect. Before, if you searched text that matched before `"*` in the Pendulum text, and after `*"` in the monster text, it would terminate on the `"` in the Pendulum text and not match, but now it keeps trying when it finds `"` again in the monster text and successfully matches.