Implement topdown duel view and option to keep the field's aspect ratio (#242)
Implement a 2d topdown duel view, at the moment the only way to toggle that is by using the f9 key or directly edit the config file. Toggling it with the f9 key will also automatically enable the option to retain the field's aspect ratio (as a stretched topdown view isn't very pleasing). Like for the topdown option the aspect ratio toggle for now can only be properly set by manually editing the config file if the keyboard function keys are not available.