Simplify unicode file handling on mingw
Add ``FileStream`` helper class that eases the creation file streams on mingw using stdio filebuf so that it can work as drop in replacement for normal fstream (on visual studio it's just an alias to fstream). This allows removing all the redoundant boilerplate needed to manually initialize streams for mingw from _wopen and co.