Fix two double free in CGUITTFont
Freetype might dealllocate the stream we provided them on its own (for example when loading a woff font), in which case when the destructor for SGUITTFace was called, we could've ended up calling delete on memory already deleted by freetype, now free the memory in the close callback.
Because of this, now if FT_Open_Face fails, we no longer have to manually free that stream either as freetype would call its close function for us.