Don't set the def string to - if the monster is a link
Fixed issues where an extra deck monster with 0 def was swapped in a tag with a link monster in its same sequence, and the def string would've remained ```-```.
This was also redoundant as in the drawing routines the cards are already checking to be links and if that's the case this string is unused.