Add support for RUSH Ritual monsters (#280)
In RUSH rituals are to be played from the Extra Deck instead of the hand. Add handling for such behavior:
In deck edit allow any ritual to be placed in either main or extra if deck edit limitations are ignored, by default rush rituals will still go to the extra and non rush will go to the main deck.
For hosting, honor the DUEL_EXTRA_DECK_RITUAL flag, and either have all the rituals go in the main or go in the extra accordingly

---------

Co-authored-by: Edoardo Lolletti <edoardo762@gmail.com>